java: Implement Query object (#39)
Implements the query object for Java. Will be used by the JUnit XML
Formatter and TestNG XML formatters and other Java plugins that switch
to using messages protocol.

Ultimately the goal replace Cucumber-JVMs own plugin system a message
based on. But this requires some tooling as messages are not easy to
work with.