Include pickle name if parameterized (#44)
Cucumber originally only had Features with Scenarios. When Scenario
Outlines were added, it became hard to distinguish which example failed.
To work around this Scenario Outline may have placeholders in their
name. These placeholders are replaced when creating a pickle name. This
helps ensure that each scenario is unique.

```
Feature: Examples Tables

  Scenario Outline: Eating <eat> cucumbers
    Given there are <start> cucumbers
    When I eat <eat> cucumbers
    Then I should have <left> cucumbers

    Examples: These are passing
      | start | eat | left |
      |    12 |   5 |    7 |
```

This would be rendered as:

```
Examples Tables
  └── Eating 5 cucumbers
```

But with the addition of the Rule element and test frameworks supporting
hierarchical test structures it also becomes desirable to include the
structure of a feature file in the test name. So we would now
render:

```
Examples Tables
└── Eating cucumbers
    └──These are passing
        └── #1.1
```

And while this hierarchy is sufficient to identify a failed example, it
is not very easy to do so when there are many. So by including the
pickle name when the Scenario Outline is parameterized we can render:

```
Examples Tables
└── Eating <eat> cucumbers
    └──These are passing
        └── #1.1: Eating 5 cucumbers
```