Add more queries (#76)
* Added several new methods to find pickles, testcases, test cast started
  and test steps by `TestStepStarted` events.
* Made lineage API public.

While Converting the TeamCity plugin to the message protocol[1] I found
that.  Unlike the JUnit XML formatter this plugin also responds to test
steps and so needs queries that involve test steps.

The plugin also needs to transform the elements in the Pickle tree to
something more complex than a name so the lineage api is made public.

1. https://github.com/cucumber/cucumber-jvm/pull/3007

Co-authored-by: David Goss <david@davidgoss.co>