Refactor acceptance tests
We currently have two implementations and one work in progress. These
implementations will develop at different speeds depending on maintainer
and contributor availability.

One of the biggest blocking elements is that the suite currently is
written such that each for each sample file all methods in the query
object are test. As a result every implementation must have all query
methods.

By refactoring the test cases such that for each query method, each
sample file produces a collection of files, one for each method, the
methods that should be implemented become opt-in.